06/10/2019 | Code refactoring, added cluster folder and some small graphical improvement.
