+++ public/manifest.json
@@ -1,6 +1,6 @@
 {
-  "short_name": "React App",
-  "name": "Create React App Sample",
+  "short_name": "Archnet",
+  "name": "Archnet: know your network",
   "icons": [
     {
       "src": "favicon.ico",
