+++ src/client/cluster/mainComponent.tsx
@@ -14,8 +14,8 @@ import CssBaseline from "@material-ui/core/CssBaseline";
 /**
  * Importing components.
  */
-import Root from "../components/rootComponent";
-import NotFound from "../components/notFoundComponent";
+import Root from "./rootComponent";
+import NotFound from "../containers/notFoundComponent";
 import { configureStore } from "../store/store";
 
 /**
@@ -26,17 +26,18 @@ import { darkTheme } from "../assets/themes/darkTheme";
 
 const store = configureStore();
 
+/**
+ * Main Component of Archnet, here you can manage everything.
+ */
 class Main extends Component {
   render() {
     return (
       <ThemeProvider theme={darkTheme}>
         <CssBaseline />
-        
         <Provider store={store}>
           <Router>
             <Switch>
                 <Route exact path="/" component={Root} />
-
                 <Route component={NotFound} />
             </Switch>
           </Router>
