+++ src/client/containers/notFoundComponent.tsx
@@ -2,7 +2,7 @@
  * Importing components from libraries.
  */
 import React, { Component } from "react";
-import UndefinedRouteComponent from "./undefinedRouteComponent";
+import UndefinedRouteComponent from "../components/undefinedRouteComponent";
 
 /**
  * Defining the AppProps interface.
@@ -18,6 +18,10 @@ interface IState {
 
 }
 
+/**
+ * The NotFound Component, manages all the undefined routes
+ * and prints on the screen the 404 page.
+ */
 class NotFound extends Component<AppProps, IState> {
   render() {
     return (
